+++ apps/frontend/src/libs/components/input/input.tsx
@@ -6,7 +6,7 @@ import {
 	type FieldPath,
 	type FieldValues,
 	type IconName,
-	type RadioInputOption,
+	type InputOption,
 } from "~/libs/types/types.js";
 
 import { Button } from "../button/button.js";
@@ -22,7 +22,7 @@ type Properties<T extends FieldValues> = {
 	label: string;
 	name: FieldPath<T>;
 	onIconClick?: (() => void) | undefined;
-	options?: RadioInputOption[];
+	options?: InputOption[];
 	placeholder?: string;
 	type?: "email" | "password" | "radio" | "text";
 };
@@ -79,7 +79,13 @@ const Input = <T extends FieldValues>({
 				{isRadioWithOptions ? (
 					<div className={styles["radio-container"]}>
 						{options.map((option) => (
-							<label className={styles["radio-option"]} key={option.value}>
+							<label
+								className={getValidClassNames(
+									styles["radio-option"],
+									isFullWidth && styles["full-width"],
+								)}
+								key={option.value}
+							>
 								<input
 									checked={field.value === option.value}
 									className={styles["radio-field"]}