+++ apps/frontend/src/pages/onboarding/onboarding.tsx
@@ -134,7 +134,6 @@ const Onboarding: React.FC = () => {
 							<div className={styles["button-container"]}>
 								{!isLastQuestion && (
 									<Button
-										isFluid
 										label="BACK"
 										onClick={handlePreviousStep}
 										type="button"
@@ -142,7 +141,6 @@ const Onboarding: React.FC = () => {
 									/>
 								)}
 								<Button
-									isFluid
 									isPrimary={isValid}
 									label={isLastQuestion ? "ANALYZE" : "NEXT"}
 									type="submit"