+++ apps/mobile/src/libs/components/wheel/libs/components/components.tsx
@@ -1,2 +1,3 @@
-export { AnimatedSector } from "./animated-sector/animated-sector";
+export { AnimatedAppearingSector } from "./animated-appearing-sector/animated-appearing-sector";
+export { AnimatedPulsingSector } from "./animated-pulsing-sector/animated-pulsing-sector";
 export { LabelOnCircle } from "./label-on-circle/label-on-circle";