+++ apps/mobile/src/libs/types/slider/slider.ts
@@ -1,0 +1,8 @@
+import { type colorToGradientColors } from "~/libs/maps/maps";
+
+type SliderData = {
+	color: keyof typeof colorToGradientColors;
+	label: string;
+};
+
+export { type SliderData };